fix layer
