+++ MutaroApp/MutaroApp/AppDelegate.swift
@@ -5,21 +5,16 @@
 //  Created by minguk-kim on 2022/12/29.
 //
 
-import FirebaseSetup
 import UIKit
-import NeedleFoundation
+import MutaroApp
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
-    private(set) var rootComponent: RootComponent!
-    
     func application(
         _ application: UIApplication,
         didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
-    ) -> Bool {
-        registerProviderFactories()
-        rootComponent = RootComponent()
-        FirebaseSetup.configure()
+    ) -> Bool {        
+        MutaroApp.shared.setup()
         
         return true
     }
